codebase_genius_collete/requirements.txt
streamlit>=1.28.0
requests>=2.31.0
markdown>=3.5.0
python-dotenv>=1.0.0
httpx
jac>=0.8.10
gitpython>=3.1.43
openai
google-generativeai
pytest
fastapi
uvicorn
plotly==5.17.0
pandas==2.1.0